+++ packages/docs/docs/main/change-log.md
@@ -12,6 +12,10 @@ This page lists the changes to the IsaacScript framework.
 ## December 8th, 2023 (Unreleased)
 
 - Added the following helper functions:
+  - `getAllBosses`
+  - `getAllNonStoryBosses`
+  - `getBossIDsForStage`
+  - `getBossIDsForStageID`
   - `getBossStageIDs`
 - Added the following types:
   - `ReadonlyRecord`
@@ -20,8 +24,9 @@ This page lists the changes to the IsaacScript framework.
   - `strict-void-functions`
   - `strict-undefined-functions`
   - `no-undefined-return-type`
-- Changed the following helper functions:
-  - `getBossSet` - Now takes in a `StageID`.
+- Deleted the following helper functions:
+  - `getBossSet`
+  - `getAllBossesSet`
 
 ## December 1st, 2023